refactor: Simplify git mode handling and improve code quality
Extract duplicated git-mode handling (create-worktree, checkout-branch,
create-branch, use-worktree, use-current-state) from agent.go and
terminal.go into a shared applyGitMode helper in worktree.go.

Backend:
- Add gitModeRequest interface and applyGitMode shared helper
- Fix double git command execution in checkoutBranchIfRequested
  and createBranchIfRequested (was running command twice on error)
- Refactor createWorktreeIfRequested to reuse resolveMainRepoRoot
  instead of duplicating 20 lines of repo-root resolution
- Optimize resolveMainRepoRoot from 4 git subprocess calls to 1
- Add 5 tests for resolveMainRepoRoot (regular repo, non-git dir,
  subdirectory, linked worktree, nested worktree)

Frontend:
- Add generation counter to fetchWorktrees to prevent stale responses
- Remove redundant rootCurrentBranch signal, reuse currentBranch
- Remove trivial handleModeChange wrapper, inline setGitMode
- Extract shared BranchSelect helper to deduplicate three
  near-identical <select> blocks in GitOptions.tsx
- Split large e2e test file into 4 focused files (55a-55d)
- Extract shared e2e helpers into helpers/worktree.ts
