+++ backend/internal/worker/service/terminal.go
@@ -46,51 +46,15 @@ func registerTerminalHandlers(d *channel.Dispatcher, svc *Context) {
 			workingDir = svc.HomeDir
 		}
 
-		// Create a worktree if requested.
-		var worktreeID string
-		if r.GetCreateWorktree() {
-			finalDir, wtID, wtErr := svc.createWorktreeIfRequested(
-				workingDir, true, r.GetWorktreeBranch(), r.GetWorktreeBaseBranch(),
-			)
-			if wtErr != nil {
-				slog.Error("failed to create worktree for terminal",
-					"error", wtErr)
-				sendInternalError(sender, "failed to create worktree: "+wtErr.Error())
-				return
-			}
-			workingDir = finalDir
-			worktreeID = wtID
+		// Apply git-mode options (create-worktree, checkout-branch, etc.).
+		gm, gmErr := svc.applyGitMode(workingDir, &r)
+		if gmErr != nil {
+			slog.Error("failed to apply git mode for terminal", "error", gmErr)
+			sendInternalError(sender, gmErr.Error())
+			return
 		}
-
-		// Checkout branch if requested.
-		if branch := r.GetCheckoutBranch(); branch != "" {
-			if err := svc.checkoutBranchIfRequested(workingDir, branch); err != nil {
-				sendInternalError(sender, "failed to checkout branch: "+err.Error())
-				return
-			}
-		}
-
-		// Use existing worktree if requested.
-		if wtPath := r.GetUseWorktreePath(); wtPath != "" {
-			finalDir, wtID, err := svc.useExistingWorktreeIfRequested(workingDir, wtPath)
-			if err != nil {
-				sendInternalError(sender, "failed to use worktree: "+err.Error())
-				return
-			}
-			workingDir = finalDir
-			if wtID != "" {
-				worktreeID = wtID
-			}
-		}
-
-		// "Use current state" — if the selected dir is an already-managed worktree, register this tab.
-		if !r.GetCreateWorktree() && r.GetCheckoutBranch() == "" && r.GetUseWorktreePath() == "" {
-			if existing, err := svc.Queries.GetWorktreeByPath(bgCtx(), workingDir); err == nil {
-				if count, err := svc.Queries.CountWorktreeTabs(bgCtx(), existing.ID); err == nil && count > 0 {
-					worktreeID = existing.ID
-				}
-			}
-		}
+		workingDir = gm.WorkingDir
+		worktreeID := gm.WorktreeID
 
 		terminalID := id.Generate()
