+++ frontend/src/components/shell/GitOptions.tsx
@@ -1,4 +1,4 @@
-import type { Component } from 'solid-js'
+import type { Accessor, Component, Setter } from 'solid-js'
 import type { GitBranchEntry } from '~/generated/leapmux/v1/git_pb'
 import type { GitMode } from '~/hooks/createWorkerDialogState'
 import { generateSlug } from 'random-word-slugs'
@@ -25,8 +25,12 @@ interface GitOptionsProps {
       worktreeBranchError?: string | null
       useWorktreePath?: string
       worktreeBaseBranch?: string
+      createBranch?: string
+      createBranchError?: string | null
+      createBranchBase?: string
     },
   ) => void
+  refreshKey?: number
   onVisibilityChange?: (visible: boolean) => void
 }
 
@@ -46,16 +50,37 @@ export const GitOptions: Component<GitOptionsProps> = (props) => {
   // Branch list for switch-branch and base branch selector
   const [branches, setBranches] = createSignal<GitBranchEntry[]>([])
   const [branchesLoading, setBranchesLoading] = createSignal(false)
-  const [rootCurrentBranch, setRootCurrentBranch] = createSignal('')
   const [selectedCheckoutBranch, setSelectedCheckoutBranch] = createSignal('')
   const [selectedBaseBranch, setSelectedBaseBranch] = createSignal('')
+
+  // Create-branch mode
+  const [newBranchName, setNewBranchName] = createSignal(generateSlug(3, { format: 'kebab' }))
+  const [selectedNewBranchBase, setSelectedNewBranchBase] = createSignal('')
 
   // Worktree list for use-worktree mode
   const [worktrees, setWorktrees] = createSignal<{ path: string, branch: string, isMain: boolean }[]>([])
   const [worktreesLoading, setWorktreesLoading] = createSignal(false)
   const [selectedWorktreePath, setSelectedWorktreePath] = createSignal('')
 
-  const branchError = createMemo(() => validateBranchName(branchName()))
+  const branchExists = (name: string) =>
+    branches().some(b => b.name === name || (b.isRemote && b.name.endsWith(`/${name}`)))
+
+  const branchError = createMemo(() => {
+    const formatErr = validateBranchName(branchName())
+    if (formatErr)
+      return formatErr
+    if (branchExists(branchName()))
+      return 'A branch with this name already exists'
+    return null
+  })
+  const newBranchError = createMemo(() => {
+    const formatErr = validateBranchName(newBranchName())
+    if (formatErr)
+      return formatErr
+    if (branchExists(newBranchName()))
+      return 'A branch with this name already exists'
+    return null
+  })
 
   const showGitOptions = () => isGitRepo() && (isRepoRoot() || isWorktreeRoot())
 
@@ -82,58 +107,6 @@ export const GitOptions: Component<GitOptionsProps> = (props) => {
     { defer: true },
   ))
 
-  // Fetch git info when worker or path changes.
-  let gitInfoGeneration = 0
-  createEffect(on(() => [props.workerId, props.selectedPath] as const, async ([wid, path]) => {
-    const gen = ++gitInfoGeneration
-    if (!wid || !path) {
-      setIsGitRepo(false)
-      setIsRepoRoot(false)
-      setIsWorktreeRoot(false)
-      setIsDirty(false)
-      setGitMode('current')
-      return
-    }
-
-    setLoading(true)
-    try {
-      const resp = await workerRpc.getGitInfo(wid, {
-        workerId: wid,
-        path,
-        orgId: org.orgId(),
-      })
-      if (gen !== gitInfoGeneration)
-        return
-      setIsGitRepo(resp.isGitRepo)
-      setIsRepoRoot(resp.isRepoRoot)
-      setIsWorktreeRoot(resp.isWorktreeRoot)
-      setIsDirty(resp.isDirty)
-      setRepoRoot(resp.repoRoot)
-      setRepoDirName(resp.repoDirName)
-      setCurrentBranch(resp.currentBranch)
-      // Reset to default mode on path change.
-      setGitMode('current')
-      // Reset branch lists.
-      setBranches([])
-      setWorktrees([])
-      setSelectedCheckoutBranch('')
-      setSelectedWorktreePath('')
-    }
-    catch {
-      if (gen !== gitInfoGeneration)
-        return
-      setIsGitRepo(false)
-      setIsRepoRoot(false)
-      setIsWorktreeRoot(false)
-      setIsDirty(false)
-      setGitMode('current')
-    }
-    finally {
-      if (gen === gitInfoGeneration)
-        setLoading(false)
-    }
-  }))
-
   // Fetch branches lazily when switch-branch or create-worktree mode is selected.
   let branchGeneration = 0
 
@@ -153,10 +126,13 @@ export const GitOptions: Component<GitOptionsProps> = (props) => {
       if (gen !== branchGeneration)
         return
       setBranches(resp.branches)
-      setRootCurrentBranch(resp.currentBranch)
-      // Default the base branch to the root repo's current branch.
-      if (!selectedBaseBranch() && resp.currentBranch) {
-        setSelectedBaseBranch(resp.currentBranch)
+      // Default the base branch to the current branch.
+      const cur = currentBranch() || resp.currentBranch
+      if (!selectedBaseBranch() && cur) {
+        setSelectedBaseBranch(cur)
+      }
+      if (!selectedNewBranchBase() && cur) {
+        setSelectedNewBranchBase(cur)
       }
     }
     catch {}
@@ -166,7 +142,10 @@ export const GitOptions: Component<GitOptionsProps> = (props) => {
     }
   }
 
+  let worktreeGeneration = 0
+
   const fetchWorktrees = async () => {
+    const gen = ++worktreeGeneration
     const wid = props.workerId
     if (!wid || !props.selectedPath)
       return
@@ -178,20 +157,87 @@ export const GitOptions: Component<GitOptionsProps> = (props) => {
         path: props.selectedPath,
         orgId: org.orgId(),
       })
-      // Filter out the main working tree if it equals the selected directory.
+      if (gen !== worktreeGeneration)
+        return
+      // Only show actual worktrees (exclude the main working tree).
       const entries = resp.worktrees
-        .filter(wt => !(wt.isMain && wt.path === props.selectedPath))
+        .filter(wt => !wt.isMain)
         .map(wt => ({ path: wt.path, branch: wt.branch, isMain: wt.isMain }))
       setWorktrees(entries)
     }
     catch {}
     finally {
-      setWorktreesLoading(false)
+      if (gen === worktreeGeneration)
+        setWorktreesLoading(false)
     }
   }
 
+  // Fetch git info when worker or path changes.
+  let gitInfoGeneration = 0
+  createEffect(on(() => [props.workerId, props.selectedPath] as const, async ([wid, path]) => {
+    const gen = ++gitInfoGeneration
+    if (!wid || !path) {
+      setIsGitRepo(false)
+      setIsRepoRoot(false)
+      setIsWorktreeRoot(false)
+      setIsDirty(false)
+      setGitMode('current')
+      return
+    }
+
+    // Only show loading state on the first fetch; when switching between
+    // git repos keep the previous UI visible to avoid a flash.
+    if (!showGitOptions()) {
+      setLoading(true)
+    }
+    try {
+      const resp = await workerRpc.getGitInfo(wid, {
+        workerId: wid,
+        path,
+        orgId: org.orgId(),
+      })
+      if (gen !== gitInfoGeneration)
+        return
+      setIsGitRepo(resp.isGitRepo)
+      setIsRepoRoot(resp.isRepoRoot)
+      setIsWorktreeRoot(resp.isWorktreeRoot)
+      setIsDirty(resp.isDirty)
+      setRepoRoot(resp.repoRoot)
+      setRepoDirName(resp.repoDirName)
+      setCurrentBranch(resp.currentBranch)
+      // Reset branch lists but preserve the selected mode.
+      setBranches([])
+      setWorktrees([])
+      setSelectedCheckoutBranch('')
+      setSelectedBaseBranch('')
+      setSelectedNewBranchBase('')
+      setSelectedWorktreePath('')
+      // Re-fetch for the current mode since the lists were cleared.
+      const mode = gitMode()
+      if (mode === 'switch-branch' || mode === 'create-branch' || mode === 'create-worktree') {
+        fetchBranches()
+      }
+      if (mode === 'use-worktree') {
+        fetchWorktrees()
+      }
+    }
+    catch {
+      if (gen !== gitInfoGeneration)
+        return
+      setIsGitRepo(false)
+      setIsRepoRoot(false)
+      setIsWorktreeRoot(false)
+      setIsDirty(false)
+      setGitMode('current')
+    }
+    finally {
+      if (gen === gitInfoGeneration)
+        setLoading(false)
+    }
+  }))
+
   createEffect(on(() => gitMode(), (mode) => {
-    if ((mode === 'switch-branch' || mode === 'create-worktree') && branches().length === 0) {
+    if ((mode === 'switch-branch' || mode === 'create-branch' || mode === 'create-worktree') && branches().length === 0) {
       fetchBranches()
     }
     if (mode === 'use-worktree' && worktrees().length === 0) {
@@ -199,16 +245,30 @@ export const GitOptions: Component<GitOptionsProps> = (props) => {
     }
   }))
 
+  // Re-fetch branches/worktrees when the refresh button is clicked.
+  createEffect(on(() => props.refreshKey, () => {
+    const mode = gitMode()
+    if (mode === 'switch-branch' || mode === 'create-branch' || mode === 'create-worktree') {
+      fetchBranches()
+    }
+    if (mode === 'use-worktree') {
+      fetchWorktrees()
+    }
+  }, { defer: true }))
+
   // Notify parent when git options change.
   createEffect(on(
-    () => [gitMode(), branchName(), branchError(), selectedCheckoutBranch(), selectedWorktreePath(), selectedBaseBranch()] as const,
-    ([mode, branch, error, checkout, wtPath, baseBranch]) => {
+    () => [gitMode(), branchName(), branchError(), selectedCheckoutBranch(), selectedWorktreePath(), selectedBaseBranch(), newBranchName(), newBranchError(), selectedNewBranchBase()] as const,
+    ([mode, branch, error, checkout, wtPath, baseBranch, nbName, nbError, nbBase]) => {
       props.onGitModeChange(mode, {
         worktreeBranch: branch,
         worktreeBranchError: error,
         checkoutBranch: checkout,
         useWorktreePath: wtPath,
         worktreeBaseBranch: baseBranch,
+        createBranch: nbName,
+        createBranchError: nbError,
+        createBranchBase: nbBase,
       })
     },
   ))
@@ -217,12 +277,57 @@ export const GitOptions: Component<GitOptionsProps> = (props) => {
     setBranchName(generateSlug(3, { format: 'kebab' }))
   }
 
-  const handleModeChange = (mode: GitMode) => {
-    setGitMode(mode)
+  const randomizeNewBranch = () => {
+    setNewBranchName(generateSlug(3, { format: 'kebab' }))
   }
+
+  const BranchSelect = (selectProps: {
+    value: Accessor<string>
+    setValue: Setter<string>
+    showPrompt?: boolean
+    showCurrent?: boolean
+  }) => (
+    <select
+      value={selectProps.value()}
+      onChange={e => selectProps.setValue(e.currentTarget.value)}
+      disabled={branchesLoading()}
+    >
+      <Show when={branchesLoading()}>
+        <option value="">Loading branches...</option>
+      </Show>
+      <Show when={!branchesLoading() && branches().length === 0}>
+        <option value="">No branches found</option>
+      </Show>
+      <Show when={!branchesLoading() && branches().length > 0}>
+        <Show when={selectProps.showPrompt}>
+          <option value="">Select a branch...</option>
+        </Show>
+        <Show when={localBranches().length > 0}>
+          <optgroup label="Local">
+            <For each={localBranches()}>
+              {b => (
+                <option value={b.name}>
+                  {b.name}
+                  {selectProps.showCurrent && b.name === currentBranch() ? ' (current)' : ''}
+                </option>
+              )}
+            </For>
+          </optgroup>
+        </Show>
+        <Show when={remoteBranches().length > 0}>
+          <optgroup label="Remote">
+            <For each={remoteBranches()}>
+              {b => <option value={b.name}>{b.name}</option>}
+            </For>
+          </optgroup>
+        </Show>
+      </Show>
+    </select>
+  )
 
   return (
     <Show when={!loading() && showGitOptions()}>
+      <div class={labelRow}>&nbsp;</div>
       <div class={radioGroup}>
         {/* Use current state */}
         <label class={radioRow}>
@@ -230,7 +335,7 @@ export const GitOptions: Component<GitOptionsProps> = (props) => {
             type="radio"
             name="git-mode"
             checked={gitMode() === 'current'}
-            onChange={() => handleModeChange('current')}
+            onChange={() => setGitMode('current')}
           />
           Use current state
         </label>
@@ -249,7 +354,7 @@ export const GitOptions: Component<GitOptionsProps> = (props) => {
             type="radio"
             name="git-mode"
             checked={gitMode() === 'switch-branch'}
-            onChange={() => handleModeChange('switch-branch')}
+            onChange={() => setGitMode('switch-branch')}
           />
           Switch to branch
         </label>
@@ -260,35 +365,46 @@ export const GitOptions: Component<GitOptionsProps> = (props) => {
                 The working copy has uncommitted changes. Switching branches may fail or discard changes.
               </div>
             </Show>
-            <select
-              value={selectedCheckoutBranch()}
-              onChange={e => setSelectedCheckoutBranch(e.currentTarget.value)}
-              disabled={branchesLoading()}
-            >
-              <Show when={branchesLoading()}>
-                <option value="">Loading branches...</option>
+            <BranchSelect value={selectedCheckoutBranch} setValue={setSelectedCheckoutBranch} showPrompt />
+          </div>
+        </Show>
+
+        {/* Create new branch */}
+        <label class={radioRow}>
+          <input
+            type="radio"
+            name="git-mode"
+            checked={gitMode() === 'create-branch'}
+            onChange={() => setGitMode('create-branch')}
+          />
+          Create new branch
+        </label>
+        <Show when={gitMode() === 'create-branch'}>
+          <div class={radioSubContent}>
+            <Show when={isDirty()}>
+              <div class={warningText}>
+                The working copy has uncommitted changes. Creating a new branch will include them.
+              </div>
+            </Show>
+            <div>
+              <div class={labelRow}>
+                Branch Name
+                <RefreshButton onClick={randomizeNewBranch} title="Generate random name" />
+              </div>
+              <input
+                type="text"
+                value={newBranchName()}
+                onInput={e => setNewBranchName(e.currentTarget.value)}
+                placeholder="feature-branch"
+              />
+              <Show when={newBranchError()}>
+                <div class={errorText}>{newBranchError()}</div>
               </Show>
-              <Show when={!branchesLoading() && branches().length === 0}>
-                <option value="">No branches found</option>
-              </Show>
-              <Show when={!branchesLoading() && branches().length > 0}>
-                <option value="">Select a branch...</option>
-                <Show when={localBranches().length > 0}>
-                  <optgroup label="Local">
-                    <For each={localBranches()}>
-                      {b => <option value={b.name}>{b.name}</option>}
-                    </For>
-                  </optgroup>
-                </Show>
-                <Show when={remoteBranches().length > 0}>
-                  <optgroup label="Remote">
-                    <For each={remoteBranches()}>
-                      {b => <option value={b.name}>{b.name}</option>}
-                    </For>
-                  </optgroup>
-                </Show>
-              </Show>
-            </select>
+            </div>
+            <div>
+              <div class={labelRow}>Base Branch</div>
+              <BranchSelect value={selectedNewBranchBase} setValue={setSelectedNewBranchBase} showCurrent />
+            </div>
           </div>
         </Show>
 
@@ -298,7 +414,7 @@ export const GitOptions: Component<GitOptionsProps> = (props) => {
             type="radio"
             name="git-mode"
             checked={gitMode() === 'create-worktree'}
-            onChange={() => handleModeChange('create-worktree')}
+            onChange={() => setGitMode('create-worktree')}
           />
           Create new worktree
         </label>
@@ -326,36 +442,7 @@ export const GitOptions: Component<GitOptionsProps> = (props) => {
             </div>
             <div>
               <div class={labelRow}>Base Branch</div>
-              <select
-                value={selectedBaseBranch()}
-                onChange={e => setSelectedBaseBranch(e.currentTarget.value)}
-                disabled={branchesLoading()}
-              >
-                <Show when={branchesLoading()}>
-                  <option value="">Loading branches...</option>
-                </Show>
-                <Show when={!branchesLoading() && branches().length > 0}>
-                  <Show when={localBranches().length > 0}>
-                    <optgroup label="Local">
-                      <For each={localBranches()}>
-                        {b => (
-                          <option value={b.name}>
-                            {b.name}
-                            {b.name === rootCurrentBranch() ? ' (current)' : ''}
-                          </option>
-                        )}
-                      </For>
-                    </optgroup>
-                  </Show>
-                  <Show when={remoteBranches().length > 0}>
-                    <optgroup label="Remote">
-                      <For each={remoteBranches()}>
-                        {b => <option value={b.name}>{b.name}</option>}
-                      </For>
-                    </optgroup>
-                  </Show>
-                </Show>
-              </select>
+              <BranchSelect value={selectedBaseBranch} setValue={setSelectedBaseBranch} showCurrent />
             </div>
             <Show when={worktreePath()}>
               <div class={pathPreview}>
@@ -373,7 +460,7 @@ export const GitOptions: Component<GitOptionsProps> = (props) => {
             type="radio"
             name="git-mode"
             checked={gitMode() === 'use-worktree'}
-            onChange={() => handleModeChange('use-worktree')}
+            onChange={() => setGitMode('use-worktree')}
           />
           Use existing worktree
         </label>
