+++ frontend/src/components/shell/NewTerminalDialog.tsx
@@ -12,7 +12,7 @@ import { WorkerSelector } from '~/components/shell/WorkerSelector'
 import { createLoadingSignal } from '~/hooks/createLoadingSignal'
 import { createWorkerDialogState } from '~/hooks/createWorkerDialogState'
 import { spinner } from '~/styles/animations.css'
-import { dialogLeftPanel, dialogRightPanel, dialogTwoColumn, errorText } from '~/styles/shared.css'
+import { dialogLeftPanel, dialogRightPanel, dialogSingleColumn, dialogTopSection, dialogTwoColumn, dialogWide, errorText } from '~/styles/shared.css'
 
 interface NewTerminalDialogProps {
   workspaceId: string
@@ -78,6 +78,8 @@ export const NewTerminalDialog: Component<NewTerminalDialogProps> = (props) => {
         worktreeBranch: state.worktreeBranch(),
         worktreeBaseBranch: state.gitMode() === 'create-worktree' ? state.worktreeBaseBranch() : '',
         checkoutBranch: state.gitMode() === 'switch-branch' ? state.checkoutBranch() : '',
+        createBranch: state.gitMode() === 'create-branch' ? state.createBranch() : '',
+        createBranchBase: state.gitMode() === 'create-branch' ? state.createBranchBase() : '',
         useWorktreePath: state.gitMode() === 'use-worktree' ? state.useWorktreePath() : '',
       })
       props.onCreated(resp.terminalId, state.workerId(), state.workingDir())
@@ -111,32 +113,31 @@ export const NewTerminalDialog: Component<NewTerminalDialogProps> = (props) => {
     </label>
   )
 
-  const leftContent = () => (
-    <>
-      <WorkerSelector state={state} />
-      <DirectorySelector state={state} />
-      {shellSelector()}
-    </>
-  )
-
   return (
-    <Dialog title="New Terminal" tall onClose={() => props.onClose()}>
+    <Dialog title="New Terminal" tall class={dialogWide} onClose={() => props.onClose()}>
       <form onSubmit={handleSubmit}>
         <section>
-          <div class={state.showGitOptions() ? dialogTwoColumn : 'vstack gap-4'}>
-            <div class={state.showGitOptions() ? dialogLeftPanel : undefined}>
-              {leftContent()}
+          <div class="vstack gap-4">
+            <div class={state.showGitOptions() ? dialogTopSection : undefined}>
+              <WorkerSelector state={state} />
+              {shellSelector()}
             </div>
-            <div class={state.showGitOptions() ? dialogRightPanel : undefined}>
-              <Show when={state.workerId()}>
-                <GitOptions
-                  workerId={state.workerId()}
-                  selectedPath={state.workingDir()}
-                  homeDir={state.workerInfoStore.getHomeDir(state.workerId())}
-                  onGitModeChange={state.handleGitModeChange}
-                  onVisibilityChange={state.setShowGitOptions}
-                />
-              </Show>
+            <div class={state.showGitOptions() ? dialogTwoColumn : dialogSingleColumn}>
+              <div class={dialogLeftPanel}>
+                <DirectorySelector state={state} />
+              </div>
+              <div class={state.showGitOptions() ? dialogRightPanel : undefined}>
+                <Show when={state.workerId()}>
+                  <GitOptions
+                    workerId={state.workerId()}
+                    selectedPath={state.workingDir()}
+                    homeDir={state.workerInfoStore.getHomeDir(state.workerId())}
+                    refreshKey={state.refreshKey()}
+                    onGitModeChange={state.handleGitModeChange}
+                    onVisibilityChange={state.setShowGitOptions}
+                  />
+                </Show>
+              </div>
             </div>
           </div>
           <Show when={state.error()}>
@@ -149,7 +150,7 @@ export const NewTerminalDialog: Component<NewTerminalDialogProps> = (props) => {
           </button>
           <button
             type="submit"
-            disabled={isTerminalCreateDisabled({ submitting: submitting.loading(), workerId: state.workerId(), workingDir: state.workingDir(), shell: shell(), gitMode: state.gitMode(), worktreeBranchError: state.worktreeBranchError(), checkoutBranch: state.checkoutBranch(), useWorktreePath: state.useWorktreePath() })}
+            disabled={isTerminalCreateDisabled({ submitting: submitting.loading(), workerId: state.workerId(), workingDir: state.workingDir(), shell: shell(), gitMode: state.gitMode(), worktreeBranchError: state.worktreeBranchError(), checkoutBranch: state.checkoutBranch(), createBranchError: state.createBranchError(), useWorktreePath: state.useWorktreePath() })}
           >
             <Show when={submitting.loading()}><Icon icon={LoaderCircle} size="sm" class={spinner} /></Show>
             {submitting.loading() ? 'Creating...' : 'Create'}
