+++ frontend/src/components/workspace/NewWorkspaceDialog.tsx
@@ -16,7 +16,7 @@ import { TabType } from '~/generated/leapmux/v1/workspace_pb'
 import { createWorkerDialogState } from '~/hooks/createWorkerDialogState'
 import { sanitizeName } from '~/lib/validate'
 import { spinner } from '~/styles/animations.css'
-import { dialogLeftPanel, dialogRightPanel, dialogTwoColumn, errorText, labelRow } from '~/styles/shared.css'
+import { dialogLeftPanel, dialogRightPanel, dialogSingleColumn, dialogTopSection, dialogTwoColumn, dialogWide, errorText, labelRow } from '~/styles/shared.css'
 
 interface NewWorkspaceDialogProps {
   onCreated: (workspace: Workspace, workerId: string) => void
@@ -61,6 +61,8 @@ export const NewWorkspaceDialog: Component<NewWorkspaceDialogProps> = (props) =>
         worktreeBranch: state.worktreeBranch(),
         worktreeBaseBranch: state.gitMode() === 'create-worktree' ? state.worktreeBaseBranch() : '',
         checkoutBranch: state.gitMode() === 'switch-branch' ? state.checkoutBranch() : '',
+        createBranch: state.gitMode() === 'create-branch' ? state.createBranch() : '',
+        createBranchBase: state.gitMode() === 'create-branch' ? state.createBranchBase() : '',
         useWorktreePath: state.gitMode() === 'use-worktree' ? state.useWorktreePath() : '',
       })
 
@@ -84,46 +86,45 @@ export const NewWorkspaceDialog: Component<NewWorkspaceDialogProps> = (props) =>
     }
   }
 
-  const leftContent = () => (
-    <>
-      <WorkerSelector state={state} />
-      <div>
-        <div class={labelRow}>
-          Title
-          <RefreshButton onClick={() => setTitle(randomTitle())} title="Generate random name" />
-        </div>
-        <input
-          type="text"
-          value={title()}
-          onInput={e => setTitle(e.currentTarget.value)}
-          placeholder="New Workspace"
-        />
-        <Show when={titleError()}>
-          <div class={errorText}>{titleError()}</div>
-        </Show>
-      </div>
-      <DirectorySelector state={state} />
-    </>
-  )
-
   return (
-    <Dialog title="New Workspace" tall onClose={() => props.onClose()}>
+    <Dialog title="New Workspace" tall class={dialogWide} onClose={() => props.onClose()}>
       <form onSubmit={handleSubmit}>
         <section>
-          <div class={state.showGitOptions() ? dialogTwoColumn : 'vstack gap-4'}>
-            <div class={state.showGitOptions() ? dialogLeftPanel : undefined}>
-              {leftContent()}
-            </div>
-            <div class={state.showGitOptions() ? dialogRightPanel : undefined}>
-              <Show when={state.workerId()}>
-                <GitOptions
-                  workerId={state.workerId()}
-                  selectedPath={state.workingDir()}
-                  homeDir={state.workerInfoStore.getHomeDir(state.workerId())}
-                  onGitModeChange={state.handleGitModeChange}
-                  onVisibilityChange={state.setShowGitOptions}
+          <div class="vstack gap-4">
+            <div class={state.showGitOptions() ? dialogTopSection : undefined}>
+              <WorkerSelector state={state} />
+              <div>
+                <div class={labelRow}>
+                  Title
+                  <RefreshButton onClick={() => setTitle(randomTitle())} title="Generate random name" />
+                </div>
+                <input
+                  type="text"
+                  value={title()}
+                  onInput={e => setTitle(e.currentTarget.value)}
+                  placeholder="New Workspace"
                 />
-              </Show>
+                <Show when={titleError()}>
+                  <div class={errorText}>{titleError()}</div>
+                </Show>
+              </div>
+            </div>
+            <div class={state.showGitOptions() ? dialogTwoColumn : dialogSingleColumn}>
+              <div class={dialogLeftPanel}>
+                <DirectorySelector state={state} />
+              </div>
+              <div class={state.showGitOptions() ? dialogRightPanel : undefined}>
+                <Show when={state.workerId()}>
+                  <GitOptions
+                    workerId={state.workerId()}
+                    selectedPath={state.workingDir()}
+                    homeDir={state.workerInfoStore.getHomeDir(state.workerId())}
+                    refreshKey={state.refreshKey()}
+                    onGitModeChange={state.handleGitModeChange}
+                    onVisibilityChange={state.setShowGitOptions}
+                  />
+                </Show>
+              </div>
             </div>
           </div>
           <Show when={state.error()}>
@@ -136,7 +137,7 @@ export const NewWorkspaceDialog: Component<NewWorkspaceDialogProps> = (props) =>
           </button>
           <button
             type="submit"
-            disabled={isWorkspaceCreateDisabled({ submitting: submitting(), workerId: state.workerId(), workingDir: state.workingDir(), titleError: titleError(), gitMode: state.gitMode(), worktreeBranchError: state.worktreeBranchError(), checkoutBranch: state.checkoutBranch(), useWorktreePath: state.useWorktreePath() })}
+            disabled={isWorkspaceCreateDisabled({ submitting: submitting(), workerId: state.workerId(), workingDir: state.workingDir(), titleError: titleError(), gitMode: state.gitMode(), worktreeBranchError: state.worktreeBranchError(), checkoutBranch: state.checkoutBranch(), createBranchError: state.createBranchError(), useWorktreePath: state.useWorktreePath() })}
           >
             <Show when={submitting()}><Icon icon={LoaderCircle} size="sm" class={spinner} /></Show>
             {submitting() ? 'Creating...' : 'Create'}
